+++ bsp_tool/branches/utoplanet/__init__.py
@@ -1,0 +1,5 @@
+"""Creators of Maerchen Busters"""
+from . import merubasu
+
+
+scripts = [merubasu]